Homework 5, part 2 (00520): Accurate + Complete
A DoublyLinkedList is a linked list in which each node has 3 fields:
- the node's data (the element),
- a pointer to the previous node in the list,
- a pointer to the next node in the list.

The only exceptions are the first and last nodes:
- the first node has no previous node, so the 'previous' field should be null;
- the last node has no next node, so the 'next' field should be null.

A DoublyLinkedList should have a 'head' pointer to the list's first node and a 'tail' pointer to the list's
last node. When the list is empty, these pointers should be null.

Homework 5, part 2 Objective:
Add 8 methods to this class.
- The running time of the first 6 methods should be O(1).
- The running time of the last 2 methods should be O(n).

Notes: 10/29/24 in NB
- The assignment operator (=) performs the operations on the right first,
then it proceeds to assign that operation to the datatype on the left.
EX: String s (2) = new String("hi"); (1)
A string object with a reference to the value "hi" is instantiated first. It is then assigned to the String variable s.

Lesson learnt:
Because this is a DoublyLinkedList, you mustn't forget that there should be 2 references between 2 nodes.
In the particular scenario where you create a new node, you must link the new node with the nodes that are already present.
Examples:
- addFirst() creates a new node, that new node must be linked to head.previous before head references the new node.
    Translated in code: head = head.previous = new Node<E>(element, null, head);
    if we only did head = new Node<E>(element, null, head);
    then the new node.next would reference head, but head.previous will still be referencing null.

- addLast() creates a new node, that new node must be linked to tail.next before tail references the new node.
    Translated in code: tail = tail.next = new Node<E> (element, tail, null);
    if we only did tail = new Node<E> (element, tail, null);
    then the new node.previous would reference tail, but tail.next will still be referencing null.